requirements.txt
numpy       # >=1.26.3
scipy       # >=1.11.4
matplotlib  # >=3.8.2
cvxopt      # >=1.3.2
osqp        # >=0.6.5
tqdm        # >=4.66.4
pyyaml
mkdocs      # >=1.5.3
mkdocstrings    # >=0.27.0
mkdocstrings-python==1.12.2
mkdocs-material # >=9.5.45
mike
imageio
trimesh
seaborns
opencv-python

# numba


